+++ src/lib/components/json-schema/form/templates/FieldTemplate.tsx
@@ -20,7 +20,7 @@ export default function FieldTemplate<T = any, F = any>(
     registry,
     required,
     rawErrors = [],
-    errors,
+    // errors,
     help,
     schema,
     uiSchema,
@@ -63,7 +63,7 @@ export default function FieldTemplate<T = any, F = any>(
         isInvalid={rawErrors && rawErrors.length > 0}
       >
         {children}
-        {errors}
+        {/* {errors} */}
         {help}
       </FormControl>
     </WrapIfAdditionalTemplate>